// src/utils/models/activos/activos.ts
export class Activos{
    nomenclaturaApto: string;
    coeficiente: string;
    cuartoUtil: string;
    nomenclaturaCuartoUtil: string[];
    vehiculo: string;
    tipoDeVehiculo: tipoVehiculos[];
    descripcionVehiculo: descricionVehiculo[];
    tieneMascota: string;
    tipoDeMascota: tipoMascota[];
    descripcionMascota: descripcionMascota[];
    telefonoFijo: string;
    }

    export class tipoVehiculos{
        tipo: string;
    }

    export class descricionVehiculo{
        descripcion: string;
    } 

    export class tipoMascota{
        tipo: string;
    }

    export class descripcionMascota{
        descripcion: string;
    }